New data structures -- take 2
Move the pin handling flags to the Controllers[] array, rather than the
Pins[] array, so they can be changed for different patches.  This
requires tracking the last value sent for each controller or event in
the LastValue[] array, in order to determine the new value for a
latching event.  Tracking values like this also lets us avoid sending
duplicate controller values, which is handy for noisy analog inputs.