Debug patches for 4ZonesBJJ, organ
Get the MIDI channels for the layers sorted out, and the organ DSP
parameters.  Add the 2nd LED, adjust the pin assignment so switches and
pins line up on my breadboard.  Basically working!

Still want a better way to initialize controller values and LED values
when switching to a new patch.