Add ability to initialize latching state
Split the "Latching" pin handling into "LatchingOff" (defaults to "off")
and "LatchingOn" (defaults to "on".)  Tweak FindProgram() to initialize
LastValue[] accordingly.  Also have it call GenerateEvent() with a
lastValue of 0x80, rather than 0xff, which makes it easier to initialize
LEDs to 0xff (fully on.)

Update the DefaultEventList[] to initialize the rotating speaker effect
on the organ patch, and its corresponding LED, to "on".  Always
initialize all LEDs in all patches to clear any leftover state from the
previous patch.

I believe everything counts as "working" in this version.  Yay!