Adjust pins to match hardware
Adjust a few pin setting to match how I ended up wiring the final
hardware. The analog pedal reads the lowest pin value when I want the
highest controller output, so adjust the code to handle that case.  Use
the constrain() and map() functions to simplify the analog input code a
bit.