Support sending events on program exit
Add EXIT_PIN, to indicate events to send before switching to a new
program. The idea is that this can undo controller changes which may
have been made on the old program, but may not be appropriate for the
new program.  On my keyboard, examples include volume and expression:
each program assumes they are at their max, and changes them from there.

(TBD if this is the best way to handle this problem... the instrument
has already switched patches and done its own initialization by the time
the EXIT_PIN events are sent.  On the other hand, I don't want to burden
each patch's INIT_PINs with undoing everything which may have been done
by any other patch....)