Fix a bug matching the bank number on the default program
Make sure the default program in the event list has a valid (handled)
bank number, so it's not skipped.

Fix a bug in FindProgram(): it should use the bank number passed in its
parameters, rather than the global LastBank variable, so it matches the
correct program when called from setup().